rename archive css to ArchivePage to maintain consistency
